refactor: move multiplication
