+++ std/math/emulated/field_ops.go
@@ -3,8 +3,6 @@ package emulated
 import (
 	"errors"
 	"fmt"
-	"math"
-	"math/big"
 
 	"github.com/consensys/gnark/frontend"
 )
@@ -97,122 +95,7 @@ func (f *Field[T]) add(a, b *Element[T], nextOverflow uint) *Element[T] {
 	return f.newInternalElement(limbs, nextOverflow)
 }
 
-// Mul computes a*b and returns it. It doesn't reduce the output and it may be
-// larger than the modulus. The returned Element has as many limbs as the inputs
-// together. If the result wouldn't fit into Element, then locally reduces the
-// inputs first. Doesn't mutate inputs.
-//
-// Even though this method skips reduction and allows for multiplication chains,
-// then in most cases it is more efficient to use [Field[T].MulMod] as reducing
-// Element with 2 times the limbs is 2 times more expensive.
-//
-// For multiplying by a constant, use [Field[T].MulConst] method which is more
-// efficient.
-//
-// Uses [MultiplicationHint].
-func (f *Field[T]) Mul(a, b *Element[T]) *Element[T] {
-	return f.reduceAndOp(f.mul, f.mulPreCond, a, b)
-}
-
-// Mul computes a*b and reduces it modulo the field order. The returned Element
-// has default number of limbs and zero overflow.
-func (f *Field[T]) MulMod(a, b *Element[T]) *Element[T] {
-	r := f.Mul(a, b)
-	return f.Reduce(r)
-}
-
-// MulConst multiplies a by a constant c and returns it. We assume that the
-// input constant is "small", so that we can compute the product by multiplying
-// all individual limbs with the constant. If it is not small, then use the
-// general [Field[T].Mul] or [Field[T].MulMod] with creating new Element from
-// the constant on-the-fly.
-func (f *Field[T]) MulConst(a *Element[T], c *big.Int) *Element[T] {
-	switch c.Sign() {
-	case -1:
-		f.MulConst(f.Neg(a), new(big.Int).Neg(c))
-	case 0:
-		return f.Zero()
-	}
-	cbl := uint(c.BitLen())
-	if cbl > f.maxOverflow() {
-		panic(fmt.Sprintf("constant bit length %d exceeds max %d", cbl, f.maxOverflow()))
-	}
-	return f.reduceAndOp(
-		func(a, _ *Element[T], u uint) *Element[T] {
-			if ba, aConst := f.constantValue(a); aConst {
-				ba.Mul(ba, c)
-				return newConstElement[T](ba)
-			}
-			limbs := make([]frontend.Variable, len(a.Limbs))
-			for i := range a.Limbs {
-				limbs[i] = f.api.Mul(a.Limbs[i], c)
-			}
-			return f.newInternalElement(limbs, a.overflow+cbl)
-		},
-		func(a, _ *Element[T]) (nextOverflow uint, err error) {
-			nextOverflow = a.overflow + uint(cbl)
-			if nextOverflow > f.maxOverflow() {
-				err = overflowError{op: "mulConst", nextOverflow: nextOverflow, maxOverflow: f.maxOverflow()}
-			}
-			return
-		},
-		a, nil,
-	)
-}
-
-func (f *Field[T]) mulPreCond(a, b *Element[T]) (nextOverflow uint, err error) {
-	reduceRight := a.overflow < b.overflow
-	nbResLimbs := nbMultiplicationResLimbs(len(a.Limbs), len(b.Limbs))
-	nextOverflow = f.fParams.BitsPerLimb() + uint(math.Log2(float64(2*nbResLimbs-1))) + 1 + a.overflow + b.overflow
-	if nextOverflow > f.maxOverflow() {
-		err = overflowError{op: "mul", nextOverflow: nextOverflow, maxOverflow: f.maxOverflow(), reduceRight: reduceRight}
-	}
-	return
-}
-
-func (f *Field[T]) mul(a, b *Element[T], nextOverflow uint) *Element[T] {
-	ba, aConst := f.constantValue(a)
-	bb, bConst := f.constantValue(b)
-	if aConst && bConst {
-		ba.Mul(ba, bb).Mod(ba, f.fParams.Modulus())
-		return newConstElement[T](ba)
-	}
-
-	// mulResult contains the result (out of circuit) of a * b school book multiplication
-	// len(mulResult) == len(a) + len(b) - 1
-	mulResult, err := f.computeMultiplicationHint(a.Limbs, b.Limbs)
-	if err != nil {
-		panic(fmt.Sprintf("multiplication hint: %s", err))
-	}
-
-	// we computed the result of the mul outside the circuit (mulResult)
-	// and we want to constrain inside the circuit that this injected value
-	// actually matches the in-circuit a * b values
-	// create constraints (\sum_{i=0}^{m-1} a_i c^i) * (\sum_{i=0}^{m-1} b_i
-	// c^i) = (\sum_{i=0}^{2m-2} z_i c^i) for c \in {1, 2m-1}
-	w := new(big.Int)
-	for c := 1; c <= len(mulResult); c++ {
-		w.SetInt64(1) // c^i
-		l := f.api.Mul(a.Limbs[0], 1)
-		r := f.api.Mul(b.Limbs[0], 1)
-		o := f.api.Mul(mulResult[0], 1)
-
-		for i := 1; i < len(mulResult); i++ {
-			w.Lsh(w, uint(c))
-			if i < len(a.Limbs) {
-				l = f.api.MulAcc(l, a.Limbs[i], w)
-			}
-			if i < len(b.Limbs) {
-				r = f.api.MulAcc(r, b.Limbs[i], w)
-			}
-			o = f.api.MulAcc(o, mulResult[i], w)
-		}
-		f.api.AssertIsEqual(f.api.Mul(l, r), o)
-	}
-	return f.newInternalElement(mulResult, nextOverflow)
-}
-
-// Reduce reduces a modulo the field order and returns it. Uses hint [RemHint].
+// Reduce reduces a modulo the field order and returns it.
 func (f *Field[T]) Reduce(a *Element[T]) *Element[T] {
 	f.enforceWidthConditional(a)
 	if a.overflow == 0 {
